evaluation package refactored
