+++ src/main/java/com/configcat/ConfigurationProvider.java
@@ -5,6 +5,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
+import com.configcat.evaluation.EvaluationDetails;
 import com.configcat.hooks.ConfigCatHooks;
 import java9.util.concurrent.CompletableFuture;
